Move AsyncClientNotProvidedError to a separate exceptions module
