+++ llama-index-integrations/vector_stores/llama-index-vector-stores-weaviate/llama_index/vector_stores/weaviate/__init__.py
@@ -1,6 +1,4 @@
-from llama_index.vector_stores.weaviate.base import (
-    WeaviateVectorStore,
-    AsyncClientNotProvidedError,
-)
+from llama_index.vector_stores.weaviate.base import WeaviateVectorStore
+from llama_index.vector_stores.weaviate._exceptions import AsyncClientNotProvidedError
 
 __all__ = ["WeaviateVectorStore", "AsyncClientNotProvidedError"]
