+++ llama-index-integrations/vector_stores/llama-index-vector-stores-weaviate/llama_index/vector_stores/weaviate/base.py
@@ -27,6 +27,7 @@ from llama_index.vector_stores.weaviate.utils import (
     get_node_similarity,
     to_node,
 )
+from llama_index.vector_stores.weaviate._exceptions import AsyncClientNotProvidedError
 
 import weaviate
 import weaviate.classes as wvc
@@ -99,17 +100,6 @@ def _to_weaviate_filter(
         return filters_list[0]
 
     return condition(filters_list)
-
-
-class AsyncClientNotProvidedError(Exception):
-    """Exception raised when the async weaviate client was not provided via  the `weaviate_aclient` parameter."""
-
-    def __init__(
-        self,
-        message="Async method called without WeaviateAsyncClient provided. Pass the async weaviate client to be used via `weaviate_aclient` to the constructor of WeaviateVectorStore.",
-    ) -> None:
-        self.message = message
-        super().__init__(self.message)
 
 
 class WeaviateVectorStore(BasePydanticVectorStore):
